refactor: Split CoreWebVitals component in separate files
