+++ frontend/src/queries/nodes/CoreWebVitals/CoreWebVitalsTab.scss
